requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
ffmpeg-python==0.2.0
future==0.18.2
fuzzywuzzy==0.18.0
idna==2.10
mp3-tagger==1.0
pytube==10.5.1
pytz==2021.1
requests==2.25.1
six==1.15.0
spotipy==2.16.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
youtube-search==2.1.0
